current updates - bug fixes for form builder and conversions.

+++ app/jobs/reprocess_item_job.rb
@@ -1,27 +1,19 @@
 # app/jobs/reprocess_item_job.rb
 
 # Rerocess Item Job
+
 class ReprocessItemJob < ApplicationJob
   queue_as :default
 
   def perform(item_id)
-    # delete converted files
-    # delete_converted_items(item_id)
-
-    working_file = Media.where(item_id: item_id, media_type: "working").first
-    ConvertingFileJob.perform_later(working_file.id)
-  end
-
-  private
-
-  # clear any converted files
-  def delete_converted_items(item_id)
-    medias = Media.where(item_id: item_id, media_type: "conversion")
-    return true if medias.empty?
-
+    # get media for Item
+    medias = Item.find(item_id).media
+    # loop through all media objects
     medias.each do |media|
-      File.delete(media.path)
-      media.destroy
+      # delete converted and thumbnail media objects
+      media.destroy if media.media_type.in?([:converted, :thumbnail])
+      # create job to process working file
+      ConvertingFileJob.perform_later(media.id) if media.media_type == :working
     end
-  end
+  end 
 end
